Separating code by functionality (Menu, Statistics, Main)
